test: remove remote usage from asar spec
